requirements.txt
wheel
flask
flask_cors
flask_script
flask_migrate
flask_sqlalchemy
gunicorn
werkzeug
pymysql
requests
pillow
pydub